Initial commit. Removed conditions_data mapper method from ProcessorV2 model. Created conditions mapping definition in the Condition model and removed method create_all_from used for importing xml.
