+++ app/models/condition.rb
@@ -1,8 +1,21 @@
 class Condition < StudyRelationship
 
-  def self.create_all_from(opts)
-    conditions = opts[:xml].xpath("//condition").collect{|xml|new(:nct_id=>opts[:nct_id],:name=>xml.text)}.flatten.compact
-    import(conditions)
+  def conditions_data
+    return unless protocol_section
+
+    nct_id = protocol_section.dig('identificationModule', 'nctId')
+
+    conditions_module = protocol_section['conditionsModule']
+    return unless conditions_module
+
+    conditions = conditions_module.dig('conditions')
+    return unless conditions
+
+    collection = []
+    conditions.each do |condition|
+      collection << { nct_id: nct_id, name: condition, downcase_name: condition.try(:downcase) }
+    end
+    collection
   end
 
 end
