+++ app/models/study_json_record/processor_v2.rb
@@ -239,24 +239,6 @@ class StudyJsonRecord::ProcessorV2
     collection
   end
 
-  def conditions_data
-    return unless protocol_section
-
-    nct_id = protocol_section.dig('identificationModule', 'nctId')
-
-    conditions_module = protocol_section['conditionsModule']
-    return unless conditions_module
-
-    conditions = conditions_module.dig('conditions')
-    return unless conditions
-
-    collection = []
-    conditions.each do |condition|
-      collection << { nct_id: nct_id, name: condition, downcase_name: condition.try(:downcase) }
-    end
-    collection
-  end
-
   def facilities_data
   end
